Separate periodic Davidson execution layers
