+++ src/mlx_atomistic/dft/_periodic_davidson.py
@@ -3,39 +3,50 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from collections.abc import Callable, Mapping, Sequence
-from dataclasses import dataclass, field
+from collections.abc import Mapping, Sequence
+from dataclasses import dataclass
 from typing import Literal
 
 import mlx.core as mx
 import numpy as np
 
 from mlx_atomistic.dft._compact import (
-    _DEFAULT_COMPACT_BATCH_POLICY,
-    _CompactBatch,
-    _CompactBatchPolicy,
     _CompactLaneState,
     _remap_initial_coefficients,
     _require_layout,
 )
+from mlx_atomistic.dft._periodic_davidson_scheduler import (
+    _DavidsonApplicationTicket,
+    _DavidsonScheduler,
+    _DavidsonScheduleResult,
+    _FixedHamiltonianToken,
+)
+from mlx_atomistic.dft._periodic_davidson_subspace import (
+    _DavidsonRitzCandidate,
+    _DavidsonRitzPair,
+    _PairedDavidsonState,
+    _projected_eigh_batch,
+    _ritz_candidate_from_projected_eigensystem,
+    _ritz_candidate_with_direct_action,
+    _ritz_pair,
+    _ritz_pair_with_direct_action,
+    _seal_ritz_candidate,
+)
 from mlx_atomistic.dft._periodic_execution import (
     _detached_failure,
     _materialize,
-    _to_numpy,
 )
 from mlx_atomistic.dft._periodic_hamiltonian import PeriodicKohnShamOperator
 from mlx_atomistic.dft._periodic_models import PeriodicDavidsonConfig, PeriodicEigenResult
 from mlx_atomistic.dft._periodic_orthonormalization import (
     _DAVIDSON_RANK_POLICY,
     _Complex64RankPolicy,
-    _RankResult,
 )
 from mlx_atomistic.dft._runtime_observer import (
     RuntimeObserver,
     add_observed_work,
     observed_phase,
 )
-from mlx_atomistic.dft.periodic_gth import PeriodicGTHNonlocalOperator
 from mlx_atomistic.dft.plane_wave import PlaneWaveBasis
 
 
@@ -50,1320 +61,6 @@ def _logical_hpsi_memory(
     fft_workspace_bytes = 2 * vector_count * grid_count * 8
     peak_temporary_bytes = fft_workspace_bytes + projector_elements * 8
     return fft_workspace_bytes, peak_temporary_bytes
-
-
-def _subspace_matrix(basis_vectors: mx.array, applied: mx.array) -> mx.array:
-    matrix = mx.conjugate(basis_vectors) @ mx.transpose(applied)
-    return 0.5 * (matrix + mx.conjugate(mx.transpose(matrix)))
-
-
-def _hamiltonian_context(
-    operator: PeriodicKohnShamOperator,
-    config: PeriodicDavidsonConfig,
-    n_bands: int,
-    rank_policy: _Complex64RankPolicy,
-) -> tuple[object, ...]:
-    nonlocal_context = (
-        None
-        if operator.nonlocal_operator is None
-        else (
-            id(operator.nonlocal_operator),
-            operator.nonlocal_operator._context_identity,
-        )
-    )
-    potential = operator._effective_local_potential
-    return (
-        id(operator),
-        id(potential),
-        tuple(int(value) for value in potential.shape),
-        str(potential.dtype),
-        operator.basis.basis_fingerprint,
-        operator.basis.order_fingerprint,
-        operator.basis._layout.lane_id,
-        operator.basis.reciprocal_grid.fingerprint,
-        tuple(float(value) for value in operator.basis.kpoint_cartesian),
-        nonlocal_context,
-        "complex64-float32",
-        str(mx.default_device()),
-        config.max_iterations,
-        config.tolerance,
-        config.max_subspace_size,
-        config.preconditioner_floor,
-        n_bands,
-        "complex64-adaptive-choleskyqr-cgs2-mgs2-rank-v6",
-        rank_policy.relative_tolerance,
-    )
-
-
-@dataclass(frozen=True, eq=False)
-class _FixedHamiltonianToken:
-    """Solve-local identity that prevents paired H(V) from crossing contexts."""
-
-    context: tuple[object, ...]
-    nonce: object = field(default_factory=object, repr=False)
-
-    @classmethod
-    def create(
-        cls,
-        operator: PeriodicKohnShamOperator,
-        config: PeriodicDavidsonConfig,
-        n_bands: int,
-        rank_policy: _Complex64RankPolicy = _DAVIDSON_RANK_POLICY,
-    ) -> _FixedHamiltonianToken:
-        return cls(_hamiltonian_context(operator, config, n_bands, rank_policy))
-
-    def validate(
-        self,
-        operator: PeriodicKohnShamOperator,
-        config: PeriodicDavidsonConfig,
-        n_bands: int,
-        rank_policy: _Complex64RankPolicy = _DAVIDSON_RANK_POLICY,
-    ) -> None:
-        if self.context != _hamiltonian_context(
-            operator,
-            config,
-            n_bands,
-            rank_policy,
-        ):
-            msg = "Davidson H(V) token does not match the fixed Hamiltonian"
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True)
-class _CompactBatchCapacity:
-    """Solve-local physical shape shared by compatible logical submissions."""
-
-    lanes: int
-    vectors: int
-    active: int
-
-
-def _finite_lane_capacity(lane_count: int, maximum: int) -> int:
-    """Return the smallest bounded power-of-two lane capacity."""
-
-    if type(lane_count) is not int or type(maximum) is not int:
-        msg = "finite lane capacities require non-bool integers"
-        raise ValueError(msg)
-    if lane_count <= 0 or maximum <= 0 or lane_count > maximum:
-        msg = "finite lane capacity must be positive and no larger than its maximum"
-        raise ValueError(msg)
-    return min(1 << (lane_count - 1).bit_length(), maximum)
-
-
-def _finite_vector_capacity(vector_count: int, maximum: int) -> int:
-    """Return a 4/8/16-style vector bucket bounded by the solve maximum."""
-
-    if type(vector_count) is not int or type(maximum) is not int:
-        msg = "finite vector capacities require non-bool integers"
-        raise ValueError(msg)
-    if vector_count <= 0 or maximum <= 0 or vector_count > maximum:
-        msg = "finite vector capacity must be positive and no larger than its maximum"
-        raise ValueError(msg)
-    capacity = min(4, maximum)
-    while capacity < vector_count:
-        capacity = min(2 * capacity, maximum)
-    return capacity
-
-
-@dataclass(frozen=True)
-class _CompactSubmission:
-    indices: tuple[int, ...]
-    capacity: _CompactBatchCapacity | None = None
-
-
-@dataclass(frozen=True)
-class _CompactSubmissionPlan:
-    submissions: tuple[_CompactSubmission, ...]
-    failures: dict[int, Exception]
-    compatibility_groups: tuple[tuple[int, ...], ...]
-
-
-class _CompactSubmissionPlanner:
-    """Build deterministic compact submissions under one bounded policy."""
-
-    def __init__(
-        self,
-        states: Sequence[_CompactLaneState],
-        *,
-        policy: _CompactBatchPolicy,
-        batch_byte_estimator: (
-            Callable[[tuple[int, ...], _CompactBatch], int] | None
-        ),
-        capacity: _CompactBatchCapacity | None,
-    ) -> None:
-        self.states = tuple(states)
-        self.policy = policy
-        self.batch_byte_estimator = batch_byte_estimator
-        self.capacity = capacity
-        self.capacity_prototype: _CompactBatch | None = None
-        self._validate_capacity()
-
-    def _validate_capacity(self) -> None:
-        capacity = self.capacity
-        if capacity is None:
-            return
-        if (
-            type(capacity.lanes) is not int
-            or type(capacity.vectors) is not int
-            or type(capacity.active) is not int
-            or capacity.lanes <= 0
-            or capacity.vectors <= 0
-            or capacity.active <= 0
-            or capacity.lanes > self.policy.batch_cap
-        ):
-            msg = "compact capacity must contain positive bounded integers"
-            raise ValueError(msg)
-
-    def plan(self) -> _CompactSubmissionPlan:
-        submissions: list[_CompactSubmission] = []
-        failures: dict[int, Exception] = {}
-        compatibility_groups: list[tuple[int, ...]] = []
-        for compatible_indices in self._compatibility_groups():
-            compatibility_groups.append(compatible_indices)
-            planned, rejected = self._plan_group(compatible_indices)
-            submissions.extend(planned)
-            failures.update(rejected)
-        return _CompactSubmissionPlan(
-            submissions=tuple(submissions),
-            failures=failures,
-            compatibility_groups=tuple(compatibility_groups),
-        )
-
-    def _compatibility_groups(self) -> tuple[tuple[int, ...], ...]:
-        grouped: dict[tuple[object, ...], list[int]] = defaultdict(list)
-        for index, state in enumerate(self.states):
-            grouped[
-                (
-                    id(state.layout.reciprocal),
-                    state.layout.grid_shape,
-                    (
-                        state.vector_count
-                        if self.capacity is None
-                        else self.capacity.vectors
-                    ),
-                )
-            ].append(index)
-        return tuple(tuple(indices) for indices in grouped.values())
-
-    def _plan_group(
-        self,
-        compatible_indices: tuple[int, ...],
-    ) -> tuple[list[_CompactSubmission], dict[int, Exception]]:
-        ordered = sorted(
-            compatible_indices,
-            key=lambda index: (
-                self.states[index].layout.active_count,
-                index,
-            ),
-        )
-        submissions: list[_CompactSubmission] = []
-        failures: dict[int, Exception] = {}
-        current: list[int] = []
-        current_batch: _CompactBatch | None = None
-        for index in ordered:
-            if len(current) == self.policy.batch_cap:
-                self._append_submission(
-                    submissions,
-                    current,
-                    current_batch,
-                    "active",
-                )
-                current = []
-                current_batch = None
-            current, current_batch = self._extend_group(
-                current,
-                current_batch,
-                index,
-                submissions,
-                failures,
-            )
-        if current:
-            self._append_submission(
-                submissions,
-                current,
-                current_batch,
-                "final",
-            )
-        return submissions, failures
-
-    def _extend_group(
-        self,
-        current: list[int],
-        current_batch: _CompactBatch | None,
-        index: int,
-        submissions: list[_CompactSubmission],
-        failures: dict[int, Exception],
-    ) -> tuple[list[int], _CompactBatch | None]:
-        candidate = [*current, index]
-        try:
-            candidate_batch = self._build_candidate(candidate)
-        except ValueError:
-            if current:
-                self._append_submission(
-                    submissions,
-                    current,
-                    current_batch,
-                    "bounded",
-                )
-            try:
-                singleton_batch = self._build_candidate((index,))
-            except ValueError as error:
-                failures[index] = _detached_failure(error)
-                return [], None
-            return [index], singleton_batch
-        return candidate, candidate_batch
-
-    def _append_submission(
-        self,
-        submissions: list[_CompactSubmission],
-        indices: Sequence[int],
-        batch: _CompactBatch | None,
-        stage: str,
-    ) -> None:
-        if batch is None:
-            msg = f"compact submission planner lost its {stage} batch"
-            raise RuntimeError(msg)
-        submissions.append(
-            _CompactSubmission(
-                tuple(indices),
-                self.capacity,
-            )
-        )
-
-    def _build_candidate(
-        self,
-        indices: Sequence[int],
-    ) -> _CompactBatch:
-        selected = [self.states[index] for index in indices]
-        capacity = self.capacity
-        if capacity is None:
-            candidate_batch = _CompactBatch.from_states(
-                selected,
-                policy=self.policy,
-            )
-        else:
-            self._validate_selected_capacity(selected)
-            if self.capacity_prototype is None:
-                self.capacity_prototype = _CompactBatch.from_states(
-                    selected[:1],
-                    policy=self.policy,
-                    lane_capacity=capacity.lanes,
-                    vector_capacity=capacity.vectors,
-                    active_capacity=capacity.active,
-                )
-            candidate_batch = self.capacity_prototype
-        estimated_bytes = (
-            candidate_batch.estimated_transient_bytes
-            if self.batch_byte_estimator is None
-            else self.batch_byte_estimator(
-                tuple(indices),
-                candidate_batch,
-            )
-        )
-        if estimated_bytes > self.policy.max_transient_bytes:
-            msg = "compact batch exceeds the complete transient byte budget"
-            raise ValueError(msg)
-        return candidate_batch
-
-    def _validate_selected_capacity(
-        self,
-        selected: Sequence[_CompactLaneState],
-    ) -> None:
-        capacity = self.capacity
-        if capacity is None:
-            raise RuntimeError("compact capacity validation requires a capacity")
-        if len(selected) > capacity.lanes:
-            msg = "compact candidate exceeds its stable lane capacity"
-            raise ValueError(msg)
-        if any(
-            state.vector_count > capacity.vectors
-            or state.layout.active_count > capacity.active
-            or (capacity.active - state.layout.active_count) / capacity.active
-            > self.policy.max_padding_fraction
-            for state in selected
-        ):
-            msg = "compact candidate exceeds its stable shape policy"
-            raise ValueError(msg)
-
-
-def _plan_compact_submissions(
-    states: Sequence[_CompactLaneState],
-    *,
-    policy: _CompactBatchPolicy,
-    batch_byte_estimator: Callable[[tuple[int, ...], _CompactBatch], int] | None = None,
-    capacity: _CompactBatchCapacity | None = None,
-) -> _CompactSubmissionPlan:
-    """Build deterministic active-count buckets within hard batch bounds."""
-
-    return _CompactSubmissionPlanner(
-        states,
-        policy=policy,
-        batch_byte_estimator=batch_byte_estimator,
-        capacity=capacity,
-    ).plan()
-
-
-def _stable_compact_capacity_groups(
-    states: Sequence[_CompactLaneState],
-    indices: Sequence[int],
-    *,
-    lane_capacity: int,
-    vector_capacity: int,
-    max_padding_fraction: float,
-) -> tuple[tuple[tuple[int, ...], _CompactBatchCapacity], ...]:
-    """Partition compatible lanes into stable active-width capacity groups."""
-
-    ordered = sorted(indices, key=lambda index: states[index].layout.active_count)
-    groups: list[tuple[tuple[int, ...], _CompactBatchCapacity]] = []
-    current: list[int] = []
-    for index in ordered:
-        candidate = [*current, index]
-        smallest = states[candidate[0]].layout.active_count
-        largest = states[candidate[-1]].layout.active_count
-        if current and (largest - smallest) / largest > max_padding_fraction:
-            active_capacity = states[current[-1]].layout.active_count
-            groups.append(
-                (
-                    tuple(current),
-                    _CompactBatchCapacity(
-                        lane_capacity,
-                        vector_capacity,
-                        active_capacity,
-                    ),
-                )
-            )
-            current = [index]
-        else:
-            current = candidate
-    if current:
-        groups.append(
-            (
-                tuple(current),
-                _CompactBatchCapacity(
-                    lane_capacity,
-                    vector_capacity,
-                    states[current[-1]].layout.active_count,
-                ),
-            )
-        )
-    return tuple(groups)
-
-
-@dataclass(frozen=True)
-class _DavidsonApplicationTicket:
-    """One lane's coefficient block awaiting a scheduled H application."""
-
-    lane_id: str
-    operator: PeriodicKohnShamOperator
-    config: PeriodicDavidsonConfig
-    n_bands: int
-    rank_policy: _Complex64RankPolicy
-    token: _FixedHamiltonianToken
-    vectors: _CompactLaneState
-    observer: RuntimeObserver | None
-    purpose: Literal["initial", "correction", "direct_validation"] = "initial"
-    capture_orbital_density: bool = False
-
-
-_DavidsonSubmissionCallback = Callable[
-    [
-        str,
-        int,
-        tuple[_DavidsonApplicationTicket, ...],
-        _CompactBatch,
-        dict[str, Exception],
-    ],
-    None,
-]
-
-
-@dataclass(frozen=True)
-class _DavidsonScheduleResult:
-    """Per-lane actions plus compatible and actually submitted groups."""
-
-    actions: dict[str, _CompactLaneState]
-    orbital_densities: dict[str, mx.array]
-    failures: dict[str, Exception]
-    groups: tuple[tuple[str, ...], ...]
-    compatibility_groups: tuple[tuple[str, ...], ...]
-
-    @property
-    def submission_count(self) -> int:
-        return len(self.groups)
-
-    def action_for(self, lane_id: str) -> _CompactLaneState:
-        failure = self.failures.get(lane_id)
-        if failure is not None:
-            raise _detached_failure(failure) from None
-        try:
-            return self.actions[lane_id]
-        except KeyError as error:
-            msg = f"Davidson scheduler has no result for lane {lane_id!r}"
-            raise ValueError(msg) from error
-
-    def orbital_density_for(self, lane_id: str) -> mx.array:
-        """Return the captured final-orbital density for one lane."""
-
-        failure = self.failures.get(lane_id)
-        if failure is not None:
-            raise _detached_failure(failure) from None
-        try:
-            return self.orbital_densities[lane_id]
-        except KeyError as error:
-            msg = f"Davidson scheduler has no orbital density for lane {lane_id!r}"
-            raise ValueError(msg) from error
-
-
-class _DavidsonScheduler:
-    """Submit compatible ragged tickets under an explicit batch-cap policy."""
-
-    def __init__(
-        self,
-        *,
-        policy: _CompactBatchPolicy = _DEFAULT_COMPACT_BATCH_POLICY,
-        submission_callback: _DavidsonSubmissionCallback | None = None,
-    ) -> None:
-        self._policy = policy
-        self._submission_callback = submission_callback
-        self._submission_index = 0
-        self._capacity_by_lane: dict[str, _CompactBatchCapacity] = {}
-
-    @property
-    def batch_cap(self) -> int:
-        return self._policy.batch_cap
-
-    @property
-    def max_padding_fraction(self) -> float:
-        return self._policy.max_padding_fraction
-
-    @property
-    def shape_policy(self) -> str:
-        return self._policy.shape_policy
-
-    def reset(self) -> None:
-        """Reset solve-local submission numbering."""
-
-        self._submission_index = 0
-        self._capacity_by_lane.clear()
-
-    @staticmethod
-    def _observer(ticket: _DavidsonApplicationTicket) -> RuntimeObserver | None:
-        return ticket.operator.observer if ticket.observer is None else ticket.observer
-
-    @staticmethod
-    def _physical_group_key(
-        ticket: _DavidsonApplicationTicket,
-    ) -> tuple[object, ...]:
-        layout = ticket.vectors.layout
-        nonlocal_operator = ticket.operator.nonlocal_operator
-        if nonlocal_operator is None:
-            nonlocal_context: object = None
-        elif isinstance(nonlocal_operator, PeriodicGTHNonlocalOperator):
-            nonlocal_context = ("gth", nonlocal_operator._context_identity)
-        else:
-            nonlocal_context = ("custom", id(nonlocal_operator))
-        return (
-            id(layout.reciprocal),
-            layout.grid_shape,
-            id(_DavidsonScheduler._observer(ticket)),
-            nonlocal_context,
-        )
-
-    def _group_key(
-        self,
-        ticket: _DavidsonApplicationTicket,
-    ) -> tuple[object, ...]:
-        capacity = self._capacity_by_lane.get(ticket.lane_id)
-        if capacity is None:
-            shape: object = ("dynamic", ticket.vectors.vector_count)
-        elif self.shape_policy == "finite-buckets":
-            shape = (
-                "finite-buckets",
-                _finite_vector_capacity(ticket.vectors.vector_count, capacity.vectors),
-                capacity.active,
-            )
-        else:
-            shape = (
-                "stable",
-                capacity.lanes,
-                capacity.vectors,
-                capacity.active,
-            )
-        return (*self._physical_group_key(ticket), shape)
-
-    def bind(self, tickets: Sequence[_DavidsonApplicationTicket]) -> None:
-        """Freeze compatible submission capacities for one Davidson solve."""
-
-        self._capacity_by_lane.clear()
-        grouped: dict[
-            tuple[object, ...],
-            list[_DavidsonApplicationTicket],
-        ] = defaultdict(list)
-        for ticket in tickets:
-            grouped[self._physical_group_key(ticket)].append(ticket)
-        for compatible in grouped.values():
-            states = [ticket.vectors for ticket in compatible]
-            vector_capacity = max(
-                max(ticket.n_bands, ticket.vectors.vector_count) for ticket in compatible
-            )
-            for indices, capacity in _stable_compact_capacity_groups(
-                states,
-                range(len(states)),
-                lane_capacity=self.batch_cap,
-                vector_capacity=vector_capacity,
-                max_padding_fraction=self.max_padding_fraction,
-            ):
-                for index in indices:
-                    self._capacity_by_lane[compatible[index].lane_id] = capacity
-
-    def apply(
-        self,
-        tickets: Sequence[_DavidsonApplicationTicket],
-    ) -> _DavidsonScheduleResult:
-        ready, failures = self._validate_tickets(tickets)
-        actions: dict[str, _CompactLaneState] = {}
-        orbital_densities: dict[str, mx.array] = {}
-        groups: list[tuple[str, ...]] = []
-        compatibility_groups: list[tuple[str, ...]] = []
-        for compatible in self._compatible_batches(ready):
-            compatibility_groups.append(
-                tuple(ticket.lane_id for ticket in compatible)
-            )
-            plan = self._plan_compatible_group(compatible)
-            for lane_index, error in plan.failures.items():
-                failures[compatible[lane_index].lane_id] = error
-            for planned in plan.submissions:
-                submission = tuple(
-                    compatible[index] for index in planned.indices
-                )
-                prepared_batch = self._prepare_submission(
-                    submission,
-                    planned,
-                )
-                lane_ids = tuple(ticket.lane_id for ticket in submission)
-                groups.append(lane_ids)
-                (
-                    submission_index,
-                    batch_actions,
-                    batch_densities,
-                    submission_failures,
-                ) = self._submit(
-                    submission,
-                    prepared_batch,
-                )
-                actions.update(batch_actions)
-                orbital_densities.update(batch_densities)
-                failures.update(submission_failures)
-                self._notify_submission(
-                    "completed",
-                    submission_index,
-                    submission,
-                    prepared_batch,
-                    submission_failures,
-                )
-        return _DavidsonScheduleResult(
-            actions=actions,
-            orbital_densities=orbital_densities,
-            failures=failures,
-            groups=tuple(groups),
-            compatibility_groups=tuple(compatibility_groups),
-        )
-
-    def _validate_tickets(
-        self,
-        tickets: Sequence[_DavidsonApplicationTicket],
-    ) -> tuple[list[_DavidsonApplicationTicket], dict[str, Exception]]:
-        if not tickets:
-            msg = "Davidson scheduler requires at least one application ticket"
-            raise ValueError(msg)
-        seen: set[str] = set()
-        failures: dict[str, Exception] = {}
-        validated: list[tuple[_DavidsonApplicationTicket, mx.array]] = []
-        for ticket in tickets:
-            if ticket.lane_id in seen:
-                msg = f"duplicate Davidson scheduler lane: {ticket.lane_id!r}"
-                raise ValueError(msg)
-            seen.add(ticket.lane_id)
-            try:
-                finite = self._validate_ticket(ticket)
-            except Exception as error:
-                failures[ticket.lane_id] = _detached_failure(error)
-            else:
-                validated.append((ticket, finite))
-        ready: list[_DavidsonApplicationTicket] = []
-        if not validated:
-            return ready, failures
-        try:
-            mx.eval(*(finite for _, finite in validated))
-        except Exception as error:
-            failure = _detached_failure(error)
-            failures.update(
-                {ticket.lane_id: failure for ticket, _ in validated}
-            )
-            return ready, failures
-        for ticket, finite in validated:
-            if bool(finite):
-                ready.append(ticket)
-            else:
-                failures[ticket.lane_id] = ValueError(
-                    "Davidson application block must be finite"
-                )
-        return ready, failures
-
-    @staticmethod
-    def _validate_ticket(ticket: _DavidsonApplicationTicket) -> mx.array:
-        if ticket.purpose not in {"initial", "correction", "direct_validation"}:
-            msg = "Davidson application purpose is invalid"
-            raise ValueError(msg)
-        if type(ticket.capture_orbital_density) is not bool or (
-            ticket.capture_orbital_density and ticket.purpose != "direct_validation"
-        ):
-            msg = "orbital density can only be captured for direct validation"
-            raise ValueError(msg)
-        ticket.token.validate(
-            ticket.operator,
-            ticket.config,
-            ticket.n_bands,
-            ticket.rank_policy,
-        )
-        ticket.operator.basis._validate_state(ticket.vectors)
-        if ticket.vectors.kind != "coefficients":
-            msg = "Davidson scheduler accepts coefficient blocks only"
-            raise ValueError(msg)
-        return mx.all(mx.isfinite(ticket.vectors.values))
-
-    def _compatible_batches(
-        self,
-        tickets: Sequence[_DavidsonApplicationTicket],
-    ) -> list[list[_DavidsonApplicationTicket]]:
-        grouped: dict[
-            tuple[object, ...],
-            list[_DavidsonApplicationTicket],
-        ] = defaultdict(list)
-        for ticket in tickets:
-            grouped[self._group_key(ticket)].append(ticket)
-        batches: list[list[_DavidsonApplicationTicket]] = []
-        for compatible in grouped.values():
-            if self.shape_policy == "finite-buckets":
-                batches.extend(
-                    compatible[start : start + self.batch_cap]
-                    for start in range(
-                        0,
-                        len(compatible),
-                        self.batch_cap,
-                    )
-                )
-            else:
-                batches.append(compatible)
-        return batches
-
-    def _plan_compatible_group(
-        self,
-        compatible: list[_DavidsonApplicationTicket],
-    ) -> _CompactSubmissionPlan:
-        def estimate_submission(
-            indices: tuple[int, ...],
-            batch: _CompactBatch,
-        ) -> int:
-            return PeriodicKohnShamOperator._estimated_batch_transient_bytes(
-                [compatible[index].operator for index in indices],
-                batch,
-                captured_density_lanes=sum(
-                    compatible[index].capture_orbital_density for index in indices
-                ),
-            )
-
-        bound_capacity = self._capacity_by_lane.get(compatible[0].lane_id)
-        if bound_capacity is not None and any(
-            self._capacity_by_lane.get(ticket.lane_id) != bound_capacity
-            for ticket in compatible
-        ):
-            msg = "Davidson stable-shape group has inconsistent capacities"
-            raise RuntimeError(msg)
-        capacity = bound_capacity
-        if capacity is not None and self.shape_policy == "finite-buckets":
-            capacity = _CompactBatchCapacity(
-                lanes=_finite_lane_capacity(
-                    len(compatible),
-                    capacity.lanes,
-                ),
-                vectors=_finite_vector_capacity(
-                    compatible[0].vectors.vector_count,
-                    capacity.vectors,
-                ),
-                active=capacity.active,
-            )
-        return _plan_compact_submissions(
-            [ticket.vectors for ticket in compatible],
-            policy=self._policy,
-            batch_byte_estimator=estimate_submission,
-            capacity=capacity,
-        )
-
-    def _prepare_submission(
-        self,
-        submission: tuple[_DavidsonApplicationTicket, ...],
-        planned: _CompactSubmission,
-    ) -> _CompactBatch:
-        prepared_batch = _CompactBatch.from_states(
-            [ticket.vectors for ticket in submission],
-            policy=self._policy,
-            lane_capacity=(
-                None if planned.capacity is None else planned.capacity.lanes
-            ),
-            vector_capacity=(
-                None if planned.capacity is None else planned.capacity.vectors
-            ),
-            active_capacity=(
-                None if planned.capacity is None else planned.capacity.active
-            ),
-        )
-        complete_transient_bytes = (
-            PeriodicKohnShamOperator._estimated_batch_transient_bytes(
-                [ticket.operator for ticket in submission],
-                prepared_batch,
-                captured_density_lanes=sum(
-                    ticket.capture_orbital_density for ticket in submission
-                ),
-            )
-        )
-        runtime_observer = self._observer(submission[0])
-        if runtime_observer is not None:
-            runtime_observer.record_peak_memory(
-                "peak_temporary_bytes",
-                complete_transient_bytes,
-            )
-        return prepared_batch
-
-    def _submit(
-        self,
-        submission: tuple[_DavidsonApplicationTicket, ...],
-        prepared_batch: _CompactBatch,
-    ) -> tuple[
-        int,
-        dict[str, _CompactLaneState],
-        dict[str, mx.array],
-        dict[str, Exception],
-    ]:
-        submission_index = self._submission_index
-        self._submission_index += 1
-        self._notify_submission(
-            "started",
-            submission_index,
-            submission,
-            prepared_batch,
-            {},
-        )
-        submission_failures: dict[str, Exception] = {}
-        accepted_actions: dict[str, _CompactLaneState] = {}
-        accepted_densities: dict[str, mx.array] = {}
-        try:
-            (
-                batch_actions,
-                batch_densities,
-                submission_failures,
-            ) = self._apply_submission(
-                submission,
-                prepared_batch,
-            )
-            for ticket in submission:
-                if (
-                    ticket.capture_orbital_density
-                    and ticket.lane_id in batch_actions
-                    and ticket.lane_id not in batch_densities
-                ):
-                    msg = "direct validation did not capture its orbital density"
-                    raise RuntimeError(msg)
-            for ticket in submission:
-                applied = batch_actions.get(ticket.lane_id)
-                if applied is None:
-                    continue
-                accepted_actions[ticket.lane_id] = applied
-                if ticket.purpose != "direct_validation":
-                    add_observed_work(
-                        ticket.observer,
-                        {
-                            "davidson_hv_new_vectors": (
-                                ticket.vectors.vector_count
-                            )
-                        },
-                    )
-                if ticket.capture_orbital_density:
-                    accepted_densities[ticket.lane_id] = batch_densities[ticket.lane_id]
-        except Exception as error:
-            failure = _detached_failure(error)
-            accepted_actions.clear()
-            accepted_densities.clear()
-            submission_failures.update(
-                {ticket.lane_id: failure for ticket in submission}
-            )
-        return (
-            submission_index,
-            accepted_actions,
-            accepted_densities,
-            submission_failures,
-        )
-
-    def _apply_submission(
-        self,
-        submission: tuple[_DavidsonApplicationTicket, ...],
-        prepared_batch: _CompactBatch,
-    ) -> tuple[
-        dict[str, _CompactLaneState],
-        dict[str, mx.array],
-        dict[str, Exception],
-    ]:
-        if len(submission) == 1 and not submission[0].capture_orbital_density:
-            ticket = submission[0]
-            applied = ticket.operator._apply_compact(
-                ticket.vectors,
-                observer=ticket.observer,
-                policy=self._policy,
-                prepared_batch=prepared_batch,
-            )
-            return {ticket.lane_id: applied}, {}, {}
-        outcome = PeriodicKohnShamOperator._apply_compact_batch(
-            tuple(ticket.operator for ticket in submission),
-            tuple(ticket.vectors for ticket in submission),
-            observer=self._observer(submission[0]),
-            policy=self._policy,
-            prepared_batch=prepared_batch,
-            capture_orbital_densities=tuple(
-                ticket.capture_orbital_density for ticket in submission
-            ),
-        )
-        actions = {
-            ticket.lane_id: outcome.actions[index]
-            for index, ticket in enumerate(submission)
-            if index in outcome.actions
-        }
-        failures = {
-            submission[index].lane_id: error
-            for index, error in outcome.failures.items()
-        }
-        orbital_densities = {
-            submission[index].lane_id: density
-            for index, density in outcome.orbital_densities.items()
-        }
-        return actions, orbital_densities, failures
-
-    def _notify_submission(
-        self,
-        status: str,
-        submission_index: int,
-        submission: tuple[_DavidsonApplicationTicket, ...],
-        prepared_batch: _CompactBatch,
-        failures: dict[str, Exception],
-    ) -> None:
-        if self._submission_callback is not None:
-            self._submission_callback(
-                status,
-                submission_index,
-                submission,
-                prepared_batch,
-                failures,
-            )
-
-
-@dataclass(frozen=True)
-class _PairedDavidsonState:
-    """Unpadded lane-local V/HV pair and its incremental projection."""
-
-    vectors: _CompactLaneState
-    applied: _CompactLaneState
-    projected: mx.array
-    token: _FixedHamiltonianToken
-
-    def __post_init__(self) -> None:
-        _require_layout(self.vectors, self.applied.layout)
-        if self.vectors.kind != "coefficients":
-            msg = "Davidson V state must contain coefficients"
-            raise ValueError(msg)
-        if self.applied.kind != "hamiltonian_action":
-            msg = "Davidson HV state must contain Hamiltonian actions"
-            raise ValueError(msg)
-        if self.vectors.vector_count != self.applied.vector_count:
-            msg = "Davidson V and HV widths must match"
-            raise ValueError(msg)
-        width = self.vectors.vector_count
-        matrix = mx.array(self.projected).astype(mx.complex64)
-        if matrix.shape != (width, width):
-            msg = "Davidson projected matrix must match the paired width"
-            raise ValueError(msg)
-        # Finiteness is checked collectively at the Rayleigh-Ritz boundary.
-        # Materializing here would serialize every lane after each append.
-        object.__setattr__(self, "projected", matrix)
-
-    @classmethod
-    def initialize(
-        cls,
-        vectors: _CompactLaneState,
-        applied: _CompactLaneState,
-        token: _FixedHamiltonianToken,
-    ) -> _PairedDavidsonState:
-        return cls(vectors, applied, _subspace_matrix(vectors.values, applied.values), token)
-
-    @property
-    def vector_count(self) -> int:
-        return self.vectors.vector_count
-
-    def require_token(self, token: _FixedHamiltonianToken) -> None:
-        if token is not self.token:
-            msg = "Davidson paired H(V) cannot cross a solve token"
-            raise ValueError(msg)
-
-    def append(
-        self,
-        vectors: _CompactLaneState,
-        applied: _CompactLaneState,
-        *,
-        token: _FixedHamiltonianToken,
-    ) -> _PairedDavidsonState:
-        self.require_token(token)
-        _require_layout(vectors, self.vectors.layout)
-        _require_layout(applied, self.vectors.layout)
-        if vectors.kind != "coefficients" or applied.kind != "hamiltonian_action":
-            msg = "Davidson append requires paired C/H(C) state"
-            raise ValueError(msg)
-        if vectors.vector_count != applied.vector_count:
-            msg = "Davidson C and H(C) widths must match"
-            raise ValueError(msg)
-        old_new = mx.conjugate(self.vectors.values) @ mx.transpose(applied.values)
-        new_new = _subspace_matrix(vectors.values, applied.values)
-        top = mx.concatenate([self.projected, old_new], axis=1)
-        bottom = mx.concatenate(
-            [mx.conjugate(mx.transpose(old_new)), new_new],
-            axis=1,
-        )
-        return _PairedDavidsonState(
-            _CompactLaneState(
-                mx.concatenate([self.vectors.values, vectors.values], axis=0),
-                self.vectors.layout,
-            ),
-            _CompactLaneState(
-                mx.concatenate([self.applied.values, applied.values], axis=0),
-                self.applied.layout,
-                "hamiltonian_action",
-            ),
-            mx.concatenate([top, bottom], axis=0),
-            token,
-        )
-
-    def transform(
-        self,
-        transform: mx.array,
-        *,
-        token: _FixedHamiltonianToken,
-    ) -> _PairedDavidsonState:
-        self.require_token(token)
-        weights = mx.array(transform).astype(mx.complex64)
-        if len(weights.shape) != 2 or int(weights.shape[1]) != self.vector_count:
-            msg = "Davidson paired transform has the wrong source width"
-            raise ValueError(msg)
-        vectors = weights @ self.vectors.values
-        applied = weights @ self.applied.values
-        projected = mx.conjugate(weights) @ self.projected @ mx.transpose(weights)
-        return _PairedDavidsonState(
-            _CompactLaneState(vectors, self.vectors.layout),
-            _CompactLaneState(
-                applied,
-                self.applied.layout,
-                "hamiltonian_action",
-            ),
-            0.5 * (projected + mx.conjugate(mx.transpose(projected))),
-            token,
-        )
-
-    def rebase_ranked(
-        self,
-        rank: _RankResult,
-        source_applied: _CompactLaneState,
-        *,
-        token: _FixedHamiltonianToken,
-    ) -> _PairedDavidsonState:
-        """Rebase H(V) onto authoritative rank-filtered coefficient values."""
-
-        self.require_token(token)
-        _require_layout(source_applied, self.vectors.layout)
-        if source_applied.kind != "hamiltonian_action":
-            msg = "Davidson ranked rebase requires Hamiltonian actions"
-            raise ValueError(msg)
-        transform = mx.array(rank.transform).astype(mx.complex64)
-        values = mx.array(rank.values).astype(mx.complex64)
-        if (
-            len(transform.shape) != 2
-            or len(values.shape) != 2
-            or int(transform.shape[0]) != int(values.shape[0])
-            or int(transform.shape[1]) != source_applied.vector_count
-        ):
-            msg = "Davidson ranked rebase transform has incompatible dimensions"
-            raise ValueError(msg)
-        vectors = _CompactLaneState(values, self.vectors.layout)
-        applied = _CompactLaneState(
-            transform @ source_applied.values,
-            self.applied.layout,
-            "hamiltonian_action",
-        )
-        return _PairedDavidsonState.initialize(vectors, applied, token)
-
-
-@dataclass(frozen=True)
-class _DavidsonRitzPair:
-    """Selected Ritz/H-Ritz values derived entirely from paired lane state."""
-
-    eigenvalues: mx.array
-    vectors: _CompactLaneState
-    applied: _CompactLaneState
-    residual_stack: mx.array
-    residuals: mx.array
-    max_residual: float
-    transform: mx.array
-
-
-@dataclass(frozen=True)
-class _DavidsonRitzCandidate:
-    """Lazy Ritz data awaiting one collective finite/residual materialization."""
-
-    eigenvalues: mx.array
-    vectors: _CompactLaneState
-    applied: _CompactLaneState
-    residual_stack: mx.array
-    residuals: mx.array
-    max_residual: mx.array
-    finite: mx.array
-    transform: mx.array
-
-
-def _ritz_residual_arrays(
-    eigenvalues: mx.array,
-    vectors: _CompactLaneState,
-    applied: _CompactLaneState,
-) -> tuple[mx.array, mx.array, mx.array, mx.array]:
-    residual_stack = applied.values - eigenvalues[:, None] * vectors.values
-    residuals = mx.sqrt(mx.sum(mx.abs(residual_stack) ** 2, axis=1))
-    max_residual = mx.max(residuals)
-    finite = mx.all(mx.isfinite(eigenvalues)) & mx.all(mx.isfinite(residuals))
-    return residual_stack, residuals, max_residual, finite
-
-
-def _ritz_residual_data(
-    eigenvalues: mx.array,
-    vectors: _CompactLaneState,
-    applied: _CompactLaneState,
-) -> tuple[mx.array, mx.array, float]:
-    residual_stack, residuals, max_residual_array, finite = _ritz_residual_arrays(
-        eigenvalues,
-        vectors,
-        applied,
-    )
-    mx.eval(max_residual_array, finite)
-    if not bool(finite):
-        msg = "Davidson Ritz data must be finite"
-        raise ValueError(msg)
-    return residual_stack, residuals, float(max_residual_array)
-
-
-def _seal_ritz_candidate(candidate: _DavidsonRitzCandidate) -> _DavidsonRitzPair:
-    if not bool(candidate.finite):
-        msg = "Davidson Ritz data must be finite"
-        raise ValueError(msg)
-    return _DavidsonRitzPair(
-        eigenvalues=candidate.eigenvalues,
-        vectors=candidate.vectors,
-        applied=candidate.applied,
-        residual_stack=candidate.residual_stack,
-        residuals=candidate.residuals,
-        max_residual=float(candidate.max_residual),
-        transform=candidate.transform,
-    )
-
-
-def _ritz_candidate_from_projected_eigensystem(
-    state: _PairedDavidsonState,
-    n_bands: int,
-    values: mx.array,
-    eigenvectors: mx.array,
-) -> _DavidsonRitzCandidate:
-    """Build lazy Ritz data from an already solved projected eigensystem."""
-
-    selected_values = mx.real(values[:n_bands])
-    selected_vectors = eigenvectors[:, :n_bands]
-    transform = mx.transpose(selected_vectors)
-    vectors = _CompactLaneState(
-        transform @ state.vectors.values,
-        state.vectors.layout,
-    )
-    applied = _CompactLaneState(
-        transform @ state.applied.values,
-        state.applied.layout,
-        "hamiltonian_action",
-    )
-    residual_stack, residuals, max_residual, finite = _ritz_residual_arrays(
-        selected_values,
-        vectors,
-        applied,
-    )
-    return _DavidsonRitzCandidate(
-        eigenvalues=selected_values,
-        vectors=vectors,
-        applied=applied,
-        residual_stack=residual_stack,
-        residuals=residuals,
-        max_residual=max_residual,
-        finite=finite,
-        transform=transform,
-    )
-
-
-def _ritz_pair_from_projected_eigensystem(
-    state: _PairedDavidsonState,
-    n_bands: int,
-    values: mx.array,
-    eigenvectors: mx.array,
-) -> _DavidsonRitzPair:
-    """Build one Ritz pair from an already solved projected eigensystem."""
-
-    candidate = _ritz_candidate_from_projected_eigensystem(
-        state,
-        n_bands,
-        values,
-        eigenvectors,
-    )
-    mx.eval(candidate.max_residual, candidate.finite)
-    return _seal_ritz_candidate(candidate)
-
-
-def _ritz_pair(state: _PairedDavidsonState, n_bands: int) -> _DavidsonRitzPair:
-    values, eigenvectors = _projected_eigh(state.projected)
-    return _ritz_pair_from_projected_eigensystem(
-        state,
-        n_bands,
-        values,
-        eigenvectors,
-    )
-
-
-def _ritz_candidate_with_direct_action(
-    candidate: _DavidsonRitzPair,
-    applied: _CompactLaneState,
-) -> _DavidsonRitzCandidate:
-    """Build lazy Ritz data whose residual uses the exact scheduled H(X)."""
-
-    _require_layout(applied, candidate.vectors.layout)
-    if applied.kind != "hamiltonian_action":
-        msg = "Davidson direct validation requires Hamiltonian actions"
-        raise ValueError(msg)
-    if applied.vector_count != candidate.vectors.vector_count:
-        msg = "Davidson direct validation width does not match its Ritz vectors"
-        raise ValueError(msg)
-    residual_stack, residuals, max_residual, finite = _ritz_residual_arrays(
-        candidate.eigenvalues,
-        candidate.vectors,
-        applied,
-    )
-    return _DavidsonRitzCandidate(
-        eigenvalues=candidate.eigenvalues,
-        vectors=candidate.vectors,
-        applied=applied,
-        residual_stack=residual_stack,
-        residuals=residuals,
-        max_residual=max_residual,
-        finite=finite,
-        transform=candidate.transform,
-    )
-
-
-def _ritz_pair_with_direct_action(
-    candidate: _DavidsonRitzPair,
-    applied: _CompactLaneState,
-) -> _DavidsonRitzPair:
-    """Return one Ritz pair whose residuals use the exact scheduled H(X)."""
-
-    direct = _ritz_candidate_with_direct_action(candidate, applied)
-    mx.eval(direct.max_residual, direct.finite)
-    return _seal_ritz_candidate(direct)
-
-
-def _projected_eigh(
-    matrix: mx.array,
-    *,
-    observer: RuntimeObserver | None = None,
-) -> tuple[mx.array, mx.array]:
-    # Only the small projected Rayleigh-Ritz matrix crosses to the CPU. LAPACK's
-    # complex128 solve avoids the complex64 convergence floor while every
-    # full-grid operator, residual, and FFT remains on the default MLX device.
-    projected = _to_numpy(matrix, dtype=np.complex128, observer=observer)
-    if projected.ndim != 2 or projected.shape[0] == 0 or projected.shape[0] != projected.shape[1]:
-        msg = "projected Rayleigh-Ritz matrix must be non-empty and square"
-        raise ValueError(msg)
-    if not np.all(np.isfinite(projected)):
-        msg = "projected Rayleigh-Ritz matrix must be finite"
-        raise ValueError(msg)
-    values, vectors = np.linalg.eigh(projected)
-    if (
-        values.shape != (projected.shape[0],)
-        or vectors.shape != projected.shape
-        or not np.all(np.isfinite(values))
-        or not np.all(np.isfinite(vectors))
-    ):
-        msg = "projected Rayleigh-Ritz eigensolve returned invalid eigenpairs"
-        raise ValueError(msg)
-    return (
-        mx.array(values.astype(np.float32)),
-        mx.array(vectors.astype(np.complex64)),
-    )
-
-
-def _projected_eigh_batch(
-    matrices: Sequence[mx.array],
-    *,
-    observer: RuntimeObserver | None = None,
-) -> tuple[tuple[mx.array, mx.array], ...]:
-    """Solve equal-width projected eigensystems through one device-to-CPU transfer."""
-
-    if not matrices:
-        msg = "projected Rayleigh-Ritz batch must be non-empty"
-        raise ValueError(msg)
-    shape = tuple(int(value) for value in matrices[0].shape)
-    if len(shape) != 2 or shape[0] == 0 or shape[0] != shape[1]:
-        msg = "projected Rayleigh-Ritz matrices must be non-empty and square"
-        raise ValueError(msg)
-    if any(tuple(int(value) for value in matrix.shape) != shape for matrix in matrices):
-        msg = "projected Rayleigh-Ritz batch matrices must have equal shapes"
-        raise ValueError(msg)
-    projected_stack = mx.stack(
-        [
-            matrix if isinstance(matrix, mx.array) else mx.array(matrix)
-            for matrix in matrices
-        ],
-        axis=0,
-    )
-    projected = _to_numpy(
-        projected_stack,
-        dtype=np.complex128,
-        observer=observer,
-    )
-    if not np.all(np.isfinite(projected)):
-        msg = "projected Rayleigh-Ritz batch must be finite"
-        raise ValueError(msg)
-    values, vectors = np.linalg.eigh(projected)
-    expected_values = (len(matrices), shape[0])
-    expected_vectors = (len(matrices), *shape)
-    if (
-        values.shape != expected_values
-        or vectors.shape != expected_vectors
-        or not np.all(np.isfinite(values))
-        or not np.all(np.isfinite(vectors))
-    ):
-        msg = "projected Rayleigh-Ritz batch returned invalid eigenpairs"
-        raise ValueError(msg)
-    return tuple(
-        (
-            mx.array(values[index].astype(np.float32)),
-            mx.array(vectors[index].astype(np.complex64)),
-        )
-        for index in range(len(matrices))
-    )
 
 
 def _initial_coefficients(basis: PlaneWaveBasis, count: int) -> _CompactLaneState:
@@ -1863,9 +560,7 @@ class _DavidsonEngine:
                     synchronize=False,
                 ):
                     paired_states = tuple(
-                        progress.paired
-                        for progress in compatible
-                        if progress.paired is not None
+                        progress.paired for progress in compatible if progress.paired is not None
                     )
                     with observed_phase(
                         observer,
@@ -1900,12 +595,10 @@ class _DavidsonEngine:
                             value
                             for _progress, candidate in candidates
                             for value in (candidate.max_residual, candidate.finite)
-                        )
+                        ),
                     )
                     for progress, candidate in candidates:
-                        pairs[progress.request.lane_id] = _seal_ritz_candidate(
-                            candidate
-                        )
+                        pairs[progress.request.lane_id] = _seal_ritz_candidate(candidate)
             except Exception:
                 # Preserve lane-local failure isolation. The single-lane path is
                 # only a recovery path for malformed or numerically failed data.
@@ -1957,9 +650,7 @@ class _DavidsonEngine:
             or progress.iteration_count >= request.config.max_iterations
         ):
             if not request.require_direct_validation:
-                progress.converged = (
-                    ritz_pair.max_residual <= request.config.tolerance
-                )
+                progress.converged = ritz_pair.max_residual <= request.config.tolerance
                 progress.done = True
                 _DavidsonEngine._emit_iteration(
                     progress,
@@ -2021,7 +712,7 @@ class _DavidsonEngine:
                         value
                         for _lane_id, candidate in candidates
                         for value in (candidate.max_residual, candidate.finite)
-                    )
+                    ),
                 )
             except Exception as error:
                 failure = _detached_failure(error)
@@ -2209,9 +900,7 @@ class _DavidsonEngine:
             try:
                 progress = self._prepare_lane(request)
                 progress_by_lane[request.lane_id] = progress
-                initial_tickets.append(
-                    self._ticket(progress, progress.initial_vectors)
-                )
+                initial_tickets.append(self._ticket(progress, progress.initial_vectors))
             except Exception as error:
                 failures[request.lane_id] = _detached_failure(error)
         if not initial_tickets:
@@ -2261,9 +950,7 @@ class _DavidsonEngine:
                     ritz_pair=ritz_pairs[lane_id],
                 )
                 if pending is not None:
-                    pending_tickets.append(
-                        self._ticket_for_pending(progress, pending)
-                    )
+                    pending_tickets.append(self._ticket_for_pending(progress, pending))
             except Exception as error:
                 self._fail_lane(progress, failures, error)
         return pending_tickets
@@ -2276,28 +963,20 @@ class _DavidsonEngine:
         summarized_observers = {
             id(progress.request.observer): progress.request.observer
             for progress in active
-            if progress.request.observer is not None
-            and not progress.request.observer.detail_events
+            if progress.request.observer is not None and not progress.request.observer.detail_events
         }
         for observer in summarized_observers.values():
             summarized_progress = [
                 progress
                 for progress in active
-                if progress.request.observer is observer
-                and progress.request.lane_id in ritz_pairs
+                if progress.request.observer is observer and progress.request.lane_id in ritz_pairs
             ]
             if not summarized_progress:
                 continue
             observer.emit(
                 "davidson_round",
-                iteration_min=min(
-                    progress.iteration_count + 1
-                    for progress in summarized_progress
-                ),
-                iteration_max=max(
-                    progress.iteration_count + 1
-                    for progress in summarized_progress
-                ),
+                iteration_min=min(progress.iteration_count + 1 for progress in summarized_progress),
+                iteration_max=max(progress.iteration_count + 1 for progress in summarized_progress),
                 active_lane_count=len(summarized_progress),
                 maximum_subspace_size=max(
                     progress.paired.vector_count
@@ -2362,10 +1041,7 @@ class _DavidsonEngine:
         followup_tickets: list[_DavidsonApplicationTicket] = []
         for ticket in tickets:
             lane_id = ticket.lane_id
-            failure = (
-                scheduled.failures.get(lane_id)
-                or direct_failures.get(lane_id)
-            )
+            failure = scheduled.failures.get(lane_id) or direct_failures.get(lane_id)
             if failure is not None:
                 self._fail_lane(
                     progress_by_lane[lane_id],
@@ -2387,9 +1063,7 @@ class _DavidsonEngine:
                     ),
                 )
                 if followup is not None:
-                    followup_tickets.append(
-                        self._ticket_for_pending(progress, followup)
-                    )
+                    followup_tickets.append(self._ticket_for_pending(progress, followup))
             except Exception as error:
                 self._fail_lane(progress, failures, error)
         return followup_tickets
